+++ vis/vis00.zig
@@ -7,7 +7,7 @@ pub fn init(_: Allocator, _: *ASCIIRay) *anyopaque {
     return @ptrFromInt(@alignOf(anyopaque));
 }
 
-pub fn step(_: *anyopaque, a: *ASCIIRay, idx: c_int) bool {
+pub fn step(_: *anyopaque, a: *ASCIIRay, idx: usize) bool {
     if (idx > 100) return true;
     a.writeln("All work and no play makes Jack a dull boy");
     return false;
